refactor: extract BaseCounter and counterTableUtils from kill/improve/delivery scripts
Introduces BaseCounter abstract class to standardize the selfPersisting guard
pattern used by counter scripts. Subclasses call setStorage()/onStorageChange()
instead of characterStorage.set/onChange directly, eliminating the need to
manage the selfPersisting flag per class. The abstract onReset() hook replaces
the repeated client.on("reset", ...) setup.

Extracts shared createPad/createHeader table-formatting helpers into
counterTableUtils.ts, removing three copies of nearly-identical code from
kill.ts, improveCounter.ts, and deliveryStats.ts.

KillCounter gains the previously missing selfPersisting guard as a side-effect
of the migration, preventing unnecessary reloads when it writes to storage.
